requirements.txt
Django==2.2.4
pytz==2019.2
sqlparse==0.3.0
psycopg2==2.8.3
appnope==0.1.0
backcall==0.1.0
decorator==4.4.0
djangorestframework==3.10.2
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
gunicorn==19.9.0
